Split `intervals.py` into multiple modules to improve maintainability and reduces unnecessary complexity.
