+++ python/tempo/intervals/core/types.py
@@ -1,0 +1,8 @@
+from datetime import datetime
+from typing import TypeVar, Union
+
+from pandas import Timestamp
+
+IntervalBoundary = Union[str, int, float, Timestamp, datetime]
+MetricValue = Union[int, float, bool]
+T = TypeVar('T')
